+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.4.2] - 2025-12-30
+
+### Added
+
+### Changed
+
+### Fixed
+- Parser: `record` as contextual keyword - works as method name, type name, field type, variable type
+
 ## [0.4.1] - 2025-12-30
 
 ### Added